chore: planning page (WIP)
